pyproject.toml
[project]
name = "talktoyoutube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.33",
    "numpy>=2.2.6",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.5.0",
    "tiktoken>=0.11.0",
    "youtube-transcript-api>=1.2.2",
]
